requirements.txt
python-telegram-bot==20.6
duckduckgo-search==8.0.2
python-dotenv==1.0.1
flask